feat(feed): combined chart + time-synced map on feed cards (#1011) + sharing polish (#1001)
Feed cards now render exactly like the activity detail view: ONE combined
multi-metric chart with per-metric toggle buttons and crosshair tooltip, plus a
time-synced interactive Leaflet map that marks the position at the hovered
chart time — on the owner's own feed and on subscribing peers' home timelines
alike, from the same structured payload.

Extraction (zero behaviour change on the detail page):
- `CombinedMetricChart` — the presentational half of EntityDetail/ActivityChart
  (D3 overlays, axis allocation, toggles, crosshair) moved to components/charts/;
  ActivityChart keeps its bucketed fetch and delegates.
- `RouteMap` — the presentational half of EntityDetail/ActivityMap (polyline +
  hover-time highlight marker) moved likewise; ActivityMap keeps its fetch.
- chart-utils + sleep-utils (and their chart/map CSS) move to components/charts/
  so components no longer live under one page's stylesheet.

Route in the structured payload:
- api-spec: `FeedStructured.route?: {t, lat, lon}[]` — present ONLY when the
  author attached the route map (`include_map`); same track and audience as the
  route.png, downsampled to ≤500 points with endpoints kept.
- Backend: `resolveStructuredRoute` (same `getLocations` source as the image)
  in the shared leaf module; both producers (public structured endpoint and the
  owner serialisation) assemble it identically — covered by a parity
  integration test asserting deep equality with an opted-in series + route.

Image suppression now per-counterpart (#1001): chart.png draws heart rate only,
so it is skipped exactly when a drawable heart-rate series renders natively;
route.png is skipped when the native map renders — in both FeedPostCard and the
home timeline (via the new pure `timelineImageVisible`).

Folded-in #1001 nits: settings resolved once per listing request (not once per
post), `buildFeedPostActivity` wires message/timeZone through, textarea capped
at the shared `feedPostMessageMaxLength`, `.timeline-structured` paragraph
spacing matches the card fallback, TZ-pinned deterministic date tests,
`structured` in alphabetical schema position, duplicated comments trimmed, MCP
`list_feed` payload divergence documented in docs/features/feed.md.

Closes #1011
Closes #1001
