+++ apps/web/src/pages/EntityDetail/style.css
@@ -286,75 +286,6 @@ a.entity-type-badge:hover {
   margin-left: auto;
 }
 
-/* Activity chart */
-.activity-chart-container {
-  position: relative;
-}
-
-.chart-toggles {
-  display: flex;
-  gap: 0.5rem;
-  margin-bottom: 0.75rem;
-  flex-wrap: wrap;
-}
-
-.chart-toggle {
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  padding: 0.25rem 0.75rem;
-  border: 1px solid #d1d5db;
-  border-radius: 999px;
-  background: transparent;
-  cursor: pointer;
-  font-size: 0.8rem;
-  color: #6b7280;
-  transition: all 0.15s ease;
-}
-
-.chart-toggle.active {
-  background: #f3f4f6;
-  border-color: #9ca3af;
-  color: #374151;
-}
-
-.chart-toggle-dot {
-  width: 8px;
-  height: 8px;
-  border-radius: 50%;
-}
-
-.chart-toggle-loading {
-  font-size: 0.8rem;
-  color: #9ca3af;
-  padding: 0.25rem 0.5rem;
-}
-
-.chart-svg-container {
-  position: relative;
-  width: 100%;
-  overflow: hidden;
-}
-
-.chart-tooltip {
-  display: none;
-  position: absolute;
-  padding: 0.375rem 0.5rem;
-  background: rgba(255, 255, 255, 0.95);
-  border: 1px solid #d1d5db;
-  border-radius: 4px;
-  font-size: 0.75rem;
-  line-height: 1.4;
-  white-space: pre;
-  pointer-events: none;
-  z-index: 10;
-  color: #374151;
-}
-
-.chart-svg-container svg {
-  display: block;
-}
-
 /* Music playlist */
 .music-playlist {
   border: 1px solid #e5e7eb;
@@ -792,31 +723,6 @@ a.entity-type-badge:hover {
   align-items: flex-start;
 }
 
-/* Activity GPS map — square up to 600px tall, never taller than wide */
-.activity-map-container {
-  width: 100%;
-  aspect-ratio: 1;
-  max-height: 600px;
-  border: 1px solid #e5e7eb;
-  border-radius: 8px;
-  overflow: hidden;
-  margin-top: 1rem;
-}
-
-.activity-map-container .leaflet-container {
-  height: 100%;
-  width: 100%;
-}
-
-.activity-map-highlight {
-  width: 14px;
-  height: 14px;
-  background: #673ab8;
-  border: 2px solid white;
-  border-radius: 50%;
-  box-shadow: 0 0 4px rgba(0, 0, 0, 0.4);
-}
-
 /* Dark mode */
 @media (prefers-color-scheme: dark) {
   .deleted-banner {
@@ -867,10 +773,6 @@ a.entity-type-badge:hover {
 
   .location-duration {
     color: #6b7280;
-  }
-
-  .activity-map-container {
-    border-color: #374151;
   }
 
   .entity-actions {
@@ -1008,23 +910,6 @@ a.entity-type-badge:hover {
 
   .edit-value-unit {
     color: #9ca3af;
-  }
-
-  .chart-toggle {
-    border-color: #4b5563;
-    color: #9ca3af;
-  }
-
-  .chart-toggle.active {
-    background: #374151;
-    border-color: #6b7280;
-    color: #d1d5db;
-  }
-
-  .chart-tooltip {
-    background: rgba(31, 41, 55, 0.95);
-    border-color: #4b5563;
-    color: #d1d5db;
   }
 }
