+++ apps/web/src/pages/Feed/TimelineStructured.css
@@ -5,6 +5,17 @@
   margin-top: 0.5rem;
 }
 
+/* The flex gap spaces the blocks; default <p> margins would double it and make
+   the title spacing differ from the FeedPostCard fallback body (#1001). */
+.timeline-structured p {
+  margin: 0;
+}
+
+/* The card's flex gap already spaces the map from the chart. */
+.timeline-structured .activity-map-container {
+  margin-top: 0;
+}
+
 .timeline-chart {
   margin: 0;
   padding: 0.5rem 0.25rem 0.25rem;
@@ -13,20 +24,9 @@
   background: #fff;
 }
 
-.timeline-chart-label {
-  font-size: 0.78rem;
-  font-weight: 600;
-  color: #374151;
-  margin: 0 0 0.25rem 0.5rem;
-}
-
 @media (prefers-color-scheme: dark) {
   .timeline-chart {
     border-color: #374151;
     background: #111827;
-  }
-
-  .timeline-chart-label {
-    color: #d1d5db;
   }
 }
